# Adapters/OpenApi/openapi.yaml
openapi: 3.0.4
info:
  title: CLI
  version: '1.0'
paths:
...
  /api/v1/equipment:
    get:
      tags:
        - Equipment
      operationId: GetAllEquipmentV1
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipmentResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipmentResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipmentResponse'
      security:
        - Bearer: [ ]
    post:
      tags:
        - Equipment
      operationId: CreateEquipmentV1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
  '/api/v1/equipment/{id}':
    get:
      tags:
        - Equipment
      operationId: GetEquipmentByIdV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    put:
      tags:
        - Equipment
      operationId: UpdateEquipmentV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    delete:
      tags:
        - Equipment
      operationId: DeleteEquipmentV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
...
  '/api/roles/{roleName}/permissions':
    get:
      tags:
        - RolePermissions
      operationId: GetRolePermissionsV1
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RolePermissionsResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionsResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/RolePermissionsResponse'
      security:
        - Bearer: [ ]
    post:
      tags:
        - RolePermissions
      operationId: AssignRolePermissionsV1
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignPermissionsToRoleRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AssignPermissionsToRoleRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AssignPermissionsToRoleRequest'
      responses:
        '200':
          description: OK
      security:
        - Bearer: [ ]
    delete:
      tags:
        - RolePermissions
      operationId: RemoveRolePermissionsV1
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemovePermissionsFromRoleRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RemovePermissionsFromRoleRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RemovePermissionsFromRoleRequest'
      responses:
        '200':
          description: OK
      security:
        - Bearer: [ ]
    put:
      tags:
        - RolePermissions
      operationId: ReplaceRolePermissionsV1
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignPermissionsToRoleRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AssignPermissionsToRoleRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AssignPermissionsToRoleRequest'
      responses:
        '200':
          description: OK
      security:
        - Bearer: [ ]
  /api/roles:
    get:
      tags:
        - Roles
      operationId: GetAllRolesV1
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RolesListResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RolesListResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/RolesListResponse'
      security:
        - Bearer: [ ]
    post:
      tags:
        - Roles
      operationId: CreateRoleV1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RoleResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
      security:
        - Bearer: [ ]
  '/api/roles/{roleName}':
    get:
      tags:
        - Roles
      operationId: GetRoleByNameV1
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RoleResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
      security:
        - Bearer: [ ]
    put:
      tags:
        - Roles
      operationId: UpdateRoleV1
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RoleResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
      security:
        - Bearer: [ ]
    delete:
      tags:
        - Roles
      operationId: DeleteRoleV1
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
      security:
        - Bearer: [ ]
...
components:
  schemas:
...
    AssignPermissionsToRoleRequest:
      required:
        - permissions
      type: object
      properties:
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/RolePermissionDto'
          nullable: true
      additionalProperties: false
...
    CreateRoleRequest:
      required:
        - color
        - description
        - displayName
        - name
      type: object
      properties:
        name:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        isDefault:
          type: boolean
      additionalProperties: false
    EquipmentRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        equipmentType:
          type: string
          nullable: true
        siteId:
          type: string
          format: uuid
      additionalProperties: false
    EquipmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        equipmentType:
          type: string
          nullable: true
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
          nullable: true
        siteId:
          type: string
          format: uuid
        siteName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    RemovePermissionsFromRoleRequest:
      required:
        - permissions
      type: object
      properties:
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/RolePermissionDto'
          nullable: true
      additionalProperties: false
...
    RoleDto:
      required:
        - color
        - description
        - displayName
        - isDefault
        - name
      type: object
      properties:
        name:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        isDefault:
          type: boolean
      additionalProperties: false
    RolePermissionDto:
      required:
        - action
        - resource
      type: object
      properties:
        resource:
          type: string
          nullable: true
        action:
          type: string
          nullable: true
      additionalProperties: false
    RolePermissionsResponse:
      required:
        - permissions
        - roleName
      type: object
      properties:
        roleName:
          type: string
          nullable: true
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/RolePermissionDto'
          nullable: true
      additionalProperties: false
    RoleResponse:
      required:
        - role
      type: object
      properties:
        role:
          $ref: '#/components/schemas/RoleDto'
      additionalProperties: false
...
    RolesListResponse:
      required:
        - roles
      type: object
      properties:
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto'
          nullable: true
      additionalProperties: false
...
    UpdateRoleRequest:
      required:
        - color
        - description
        - displayName
      type: object
      properties:
        displayName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        isDefault:
          type: boolean
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      description: Enter the JWT token
      scheme: Bearer
      bearerFormat: JWT
...